backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23,<1.0
pydantic>=2.3,<3.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
# Optional: only used if you enable real SMS sending
# twilio>=9.0,<10.0
# Optional: only used if you enable real Azure OpenAI calls
openai>=1.40.0
pyyaml>=6.0.1
